P4ADEV-2134 merge BrokerService into single class
