+++ src/main/java/it/gov/pagopa/pu/debtpositions/connector/organization/service/BrokerService.java
@@ -7,4 +7,6 @@ import java.util.Optional;
 public interface BrokerService {
 
   Optional<Broker> getBrokerByOrganizationId(Long organizationId, String accessToken);
+
+  Broker findById(Long brokerId, String accessToken);
 }
