+++ src/main/java/it/gov/pagopa/pu/debtpositions/service/ReceiptServiceImpl.java
@@ -1,7 +1,7 @@
 package it.gov.pagopa.pu.debtpositions.service;
 
-import it.gov.pagopa.pu.debtpositions.connector.organization.BrokerService;
-import it.gov.pagopa.pu.debtpositions.connector.organization.OrganizationService;
+import it.gov.pagopa.pu.debtpositions.connector.organization.service.BrokerService;
+import it.gov.pagopa.pu.debtpositions.connector.organization.service.OrganizationService;
 import it.gov.pagopa.pu.debtpositions.dto.Receipt;
 import it.gov.pagopa.pu.debtpositions.dto.generated.InstallmentStatus;
 import it.gov.pagopa.pu.debtpositions.dto.generated.ReceiptDTO;
